Inject controller config into the uniter
This starts the injection of the controller config service into
the unit agent. Unfortunately, the tests are a bit of a mess. So
I've spent sometime splitting the tests into individual files. This
should hopefully make it easier to understand where a test should be.

Also the mocks have been removed to prevent others from using other
packages mocks. They're not your mocks, they're someone elses.

There is still some work to sort out the construction of the facade,
but we can do that in another change.
