+++ domain/testing/controllerconfigsuite.go
@@ -11,11 +11,8 @@ import (
 
 	"github.com/juju/juju/controller"
 	coredatabase "github.com/juju/juju/core/database"
-	databasetesting "github.com/juju/juju/database/testing"
-	cloudbootstrap "github.com/juju/juju/domain/cloud/bootstrap"
 	"github.com/juju/juju/domain/controllerconfig/bootstrap"
 	schematesting "github.com/juju/juju/domain/schema/testing"
-	"github.com/juju/juju/juju/testing"
 )
 
 // ControllerConfigSuite is used to provide a sql.DB reference to tests.
@@ -30,7 +27,6 @@ type ControllerConfigSuite struct {
 // with the controller config.
 func (s *ControllerConfigSuite) SetUpTest(c *gc.C) {
 	s.ControllerSuite.SetUpTest(c)
-	s.SeedControllerCloud(c, s.TxnRunner())
 	s.SeedControllerConfig(c, s.TxnRunner(), s.ControllerConfig)
 }
 
@@ -38,14 +34,5 @@ func (s *ControllerConfigSuite) SetUpTest(c *gc.C) {
 // the given database.
 func (s *ControllerConfigSuite) SeedControllerConfig(c *gc.C, runner coredatabase.TxnRunner, config controller.Config) {
 	err := bootstrap.InsertInitialControllerConfig(config)(context.Background(), runner)
-	c.Assert(err, jc.ErrorIsNil)
-}
-
-// SeedControllerCloud is responsible for applying the controller cloud to
-// the given database.
-func (s *ControllerConfigSuite) SeedControllerCloud(c *gc.C, runner coredatabase.TxnRunner) {
-	err := databasetesting.DummyCloudOpt(context.Background(), s.TxnRunner())
-	c.Assert(err, jc.ErrorIsNil)
-	err = cloudbootstrap.InsertInitialControllerCloud(testing.DefaultCloud)(context.Background(), runner)
 	c.Assert(err, jc.ErrorIsNil)
 }
